pyproject.toml
[tool.poetry]
name = "llm_processes"
version = "0.0.1"
description = "LLM Processes: Numerical Predictive Distributions Conditioned on Natural Language"
authors = ["John Bronskill", "James Requeima", "Dami Choi"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/requeima/llm_processes"
repository = "https://github.com/requeima/llm_processes"

[tool.poetry.dependencies]
python = ">=3.10"
torch = ">=2.3.0"
transformers = ">=4.41.0"
accelerate = ">=0.30.1"
jsonargparse = ">=4.28.0"
matplotlib = ">=3.9.0"
llama-cpp-python = ">=0.2.79"
optuna = {version = ">=3.6.1", optional = true}
gpytorch = {version = ">=1.14", optional = true}

[tool.poetry.extras]
opt = ["optuna", "gpytorch"]

[tool.poetry.dev-dependencies]
pytest = "^6.2"
# Add other development dependencies here

[tool.poetry.scripts]
llm_process = "llm_processes.run_llm_process:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
